Add `abstract` key to the `provides` objects.
This is so that each extension can have its own description. Bump the version
to 1.0.0dev4 to reflect this change, which is backwards compatible (because
the abstract key is optional).

In passing, fix a few markdown nerbles that were making the formatting of the
HTML version slightly inconsistent. Also eliminated some unnecessary
`<strong>` and `<a>` tags in favor of the corresponding Markdown syntax, and
updated the HTML using the latest version of discount.